navbar bootstrapified
