+++ package.json
@@ -6,7 +6,10 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.7.3",
+    "bootstrap": "^4.6.0",
+    "bootswatch": "^4.6.0",
     "react": "^17.0.1",
+    "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.1",
     "react-hook-form": "^6.15.4",
     "react-router-dom": "^5.2.0",
