+++ src/components/OpenLibrary/OpenLibrary.js
@@ -1,11 +1,11 @@
 import React from "react"
-import { NavBar } from "../Nav/Nav"
+import { Navigation } from "../Nav/Nav"
 import { OpenLibrarySearch } from "./OpenLibarySearch"
 
 export const OpenLibrary = (props) => {
   return (
     <>
-      <NavBar {...props} />
+      <Navigation {...props} />
       <OpenLibrarySearch {...props} />
     </>
   )
