+++ src/index.js
@@ -2,6 +2,8 @@ import React from "react"
 import ReactDOM from "react-dom"
 import { BrowserRouter as Router } from "react-router-dom"
 import { App } from "./App"
+import "bootstrap/dist/css/bootstrap.min.css"
+import "bootswatch/dist/flatly/bootstrap.min.css"
 // import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
